Fix MSVC /W4 /WX failures in two native tests
The Windows CI job compiles the native suite with warnings as errors.
candle_bubble_history.h passed int64_t bounds to a double parameter
(C4244), and flow_positioning_test.cpp declared a global named start that
shadowed a local in nlohmann's json_pointer.hpp (C4459). Cast the bounds
explicitly and rename the test constant to t0.

CONTRIBUTING.md claimed the suite needs no network access, but the
workspace and flow positioning tests fetch nlohmann JSON on first
configure. Restore the accurate sentence and the offline override.

+++ src/core/candle_bubble_history.h
@@ -133,7 +133,7 @@ private:
             for (const auto& p : *prints) {
                 if (!p.is_object()) {fail();return;}
                 Terminal::Trade t{};
-                workspace::read(p,"timestamp_ms",t.timestamp_ms,start,start+59999);
+                workspace::read(p,"timestamp_ms",t.timestamp_ms,double(start),double(start+59999));
                 workspace::read(p,"price",t.price,0,1e18); workspace::read(p,"qty",t.qty,0,1e18);
                 if (!p.contains("is_buy") || !p["is_buy"].is_boolean() || !p.contains("id") || !p["id"].is_string()) {fail();return;}
                 t.is_buy = p["is_buy"].get<bool>();
